+++ MordoEngine/src/Scene/Editor/System/EditorSystem.h
@@ -14,11 +14,17 @@ private:
 	float m_Radius = 25.0f;
 	int m_HeightOffSet = 12.0f;
 	float m_BrushStrenght = 10.0f;
+	float m_RadiusStep = 2.0f;
+	float m_BrushStrenghtStep = 0.5f;
 	glm::vec3 m_LastWorldPosition;
 	glm::vec3 RaycastToTerrain(
 		const glm::vec3& rayOrigin,
 		const glm::vec3& rayDir,
 		const terrain::Terrain& terrain);
+	void ModyfySelector(float radius);
+	void ModifyTerrain(terrain::Terrain& terrain, float heightFactor);
+	void ModifyBrushStrenght(float strenghtFactor);
+
 public:
 	EditorSystem(Shader& areaSelectorShader);
 	void Update(const terrain::Terrain& terrain, const Camera& camera);
@@ -27,5 +33,12 @@ public:
 				glm::mat4* model,
 				glm::vec3* lightDir);
 	glm::vec3 GetWorldPosition() const;
-	void ModifyTerrain(terrain::Terrain& terrain);
+	void IncreaseSelector();
+	void DecreaseSelector();
+
+	void IncreaseTerrain(terrain::Terrain& terrain);
+	void DecreaseTerrain(terrain::Terrain& terrain);
+
+	void IncreaseBrushStrenght();
+	void DecreaseBrushStrenght();
 };